requirements.txt
# Core
sentence-transformers>=2.2.0
transformers>=4.30.0
torch>=2.0.0
python-dotenv>=1.0.0

# LLM Clients
groq>=0.9.0
google-generativeai>=0.5.0

# Text Processing & Matching
rapidfuzz>=3.0.0

# Data & Analysis
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.2.0

# Visualization (for notebook)
matplotlib>=3.7.0
seaborn>=0.12.0

# Development
pytest>=7.0.0
black>=23.0.0
